Extract some components from App.vue
